move session cleanup tasks to AuthScreens
