+++ src/libs/Navigation/AppNavigator/AuthScreens.js
@@ -70,6 +70,7 @@ import CardOverlay from '../../../components/CardOverlay';
 import defaultScreenOptions from './defaultScreenOptions';
 import * as API from '../../API';
 import {setLocale} from '../../actions/App';
+import {cleanupSession} from '../../actions/Session';
 
 Onyx.connect({
     key: ONYXKEYS.MY_PERSONAL_DETAILS,
@@ -249,7 +250,7 @@ class AuthScreens extends React.Component {
         if (this.unsubscribeGroupShortcut) {
             this.unsubscribeGroupShortcut();
         }
-        NetworkConnection.stopListeningForReconnect();
+        cleanupSession();
         clearInterval(this.interval);
         this.interval = null;
         hasLoadedPolicies = false;
