--- src/libs/SignoutManager/index.native.js
@@ -1,6 +1,0 @@
-// Native does not need to worry about cross tab communication
-
-export default {
-    notifyTabsOfSignout: () => {},
-    onSignout: () => {},
-};
