+++ src/libs/actions/Session.js
@@ -13,7 +13,10 @@ import Navigation from '../Navigation/Navigation';
 import ROUTES from '../../ROUTES';
 import {translateLocal} from '../translate';
 import * as Network from '../Network';
-
+import UnreadIndicatorUpdater from '../UnreadIndicatorUpdater';
+import Timers from '../Timers';
+import * as Pusher from '../Pusher/pusher';
+import NetworkConnection from '../NetworkConnection';
 
 let credentials = {};
 Onyx.connect({
@@ -306,6 +309,19 @@ function continueSessionFromECom(accountID, validateCode, twoFactorAuthCode) {
     });
 }
 
+/**
+ * Put any logic that needs to run when we are signed out here. This can be triggered when the current tab or another tab signs out.
+ */
+function cleanupSession() {
+    // We got signed out in this tab or another so clean up any subscriptions and timers
+    NetworkConnection.stopListeningForReconnect();
+    UnreadIndicatorUpdater.stopListeningForReportChanges();
+    PushNotification.deregister();
+    PushNotification.clearNotifications();
+    Pusher.disconnect();
+    Timers.clearAll();
+}
+
 export {
     continueSessionFromECom,
     fetchAccountDetails,
@@ -315,4 +331,5 @@ export {
     reopenAccount,
     resendValidationLink,
     resetPassword,
+    cleanupSession,
 };
